requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
docopt==0.6.2
entrypoints==0.4
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0
fonttools==4.29.1
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.3.3
# git+https://github.com/repodiac/german_transliterate
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==2.10.0
huggingface-hub==0.0.8
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
inflect==5.4.0
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jamo==0.4.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
num2words==0.5.10
numba==0.55.1
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.2
pooch==1.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
# pyopenjtalk
pyparsing==3.0.7
pypinyin==0.46.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyworld==0.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.4.1
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-io-gcs-filesystem==0.24.0
# TensorFlowTTS==0.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
TextGrid==1.5
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.3
typing_extensions==4.1.1
Unidecode==1.3.3
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
